Change lesson 4 intro to text view - adding a new fragment for the intro
